Relax sweeping for implicit teams to avoid invalid sigs
Signature chains like:
sig 1 - change_membership, remove UV
sig 2 - add keybase-invite

are invalid for implicit team, because as of sig 1 we are entirely
removing member of a team - crypto memberships have to be constant
at every sig in implicit team.

So what we do now is a compat-mode for implicit team that will allow
team transaction to add a PUKless member type invite that would
duplicate existing crypto dude.